+++ src/pages/data/components/BrowseDataTable.tsx
@@ -1,16 +1,12 @@
-import { Buffer } from 'buffer';
 import { FreeTagIcon } from '@assets/icons';
 import CustomButton from '@components/ui/Button';
 import CustomTable, { ITableColumns } from '@components/ui/CustomTable';
 import PriceButton from '@components/ui/PriceButton';
 import TruncatedAddress from '@shared/components/TruncatedAddress';
-import {Link, useNavigate} from 'react-router-dom';
-import { createTransaction, sendTransaction } from '@slices/transactionSlice';
+import {Link} from 'react-router-dom';
 import { useAuth } from '@contexts/auth-provider';
-import { VersionedTransaction } from '@solana/web3.js';
-import { useWallet } from '@solana/wallet-adapter-react';
 import useDownloadDataset from '../hooks/useDownloadDataset';
-import { IDataset } from '@store/data/types';
+import useTransaction from '@shared/hooks/useTransaction';
 
 const BrowseDataTable = ({
   data,
@@ -19,36 +15,10 @@ const BrowseDataTable = ({
   data: Record<string, any>[];
   isLoading: boolean;
 }) => {
+  console.log(data)
   const { address } = useAuth();
-  const { signTransaction } = useWallet();
-  const navigate = useNavigate();
+  const { handlePurchase } = useTransaction();
   const { handleDownload, isLoading : isDownloading } = useDownloadDataset();
-
-  const handlePurchase = async (
-    dataset: IDataset
-  ) => {
-    if (!dataset.item_hash) throw new Error('Item hash is undefined');
-    if (!signTransaction) throw new Error('Wallet not connected');
-
-    const { transaction } = await createTransaction({
-      signer: address,
-      datasetId: dataset.item_hash,
-    })
-
-    const serializedBuffer = Buffer.from(transaction, 'base64');
-    const unsignedTransaction = VersionedTransaction.deserialize(serializedBuffer);
-    const signedTransaction = await signTransaction(unsignedTransaction);
-    const signedSerializedBase64 = Buffer.from(signedTransaction.serialize()).toString('base64');
-    const response = await sendTransaction({
-        transaction: signedSerializedBase64,
-        datasetId: dataset.item_hash,
-    });
-
-    console.log(response);
-
-    navigate(`/data/${dataset.item_hash}`);
-  };
-
   const COLUMNS: ITableColumns[] = [
     {
       header: 'NAME',
@@ -85,7 +55,6 @@ const BrowseDataTable = ({
             <p className="w-[120px]">Free</p>
           </div>
         ) : (
-          // make this a component so you can pass use the transaction hook that you wrote in it
           <PriceButton price={price} />
         ),
       sortWith: 'price',
@@ -99,7 +68,7 @@ const BrowseDataTable = ({
       cell: (dataset) => (
         <div className="w-auto flex items-end justify-end">
           {/* eslint-disable-next-line no-nested-ternary */}
-          {dataset.available && dataset.price == 0 || dataset.permission_status === 'GRANTED' ? (
+          {dataset.price === '0' || dataset?.permission_status === 'GRANTED' ? (
             <CustomButton
               text="Download"
               btnStyle="outline-primary"
@@ -130,7 +99,7 @@ const BrowseDataTable = ({
               size="sm"
               icon="buy"
               btnStyle="solid-secondary"
-              onClick={() => handlePurchase(dataset)}
+              onClick={() => handlePurchase(dataset.item_hash)}
             />
           )}
         </div>
