+++ src/shared/hooks/useTransaction.ts
@@ -1,16 +1,38 @@
+import { createTransaction, sendTransaction } from '@slices/transactionSlice';
+import { useAuth } from '@contexts/auth-provider';
 import { useWallet } from '@solana/wallet-adapter-react';
-import { useState } from 'react';
-import { parseEther } from 'viem';
-import { usePrepareSendTransaction, useSendTransaction } from 'wagmi';
+import { VersionedTransaction } from '@solana/web3.js';
+import { useCallback } from 'react';
+import { useNavigate } from 'react-router-dom';
+import { Buffer } from 'buffer';
 
 export default () => {
-  // eslint-disable-next-line @typescript-eslint/naming-convention
-  // const { owner, price } = useAppSelector(
-  //   (state) => state.datasets.dataDetails
-  // );
-  const [dataPrice, setPrice] = useState<string>('0');
-  const [recipientAddress, setRecipientAddress] =
-    useState<string>('hashtagggsomething');
-  const { sendTransaction } = useWallet();
-  return { setPrice, setRecipientAddress, sendTransaction };
+  const navigate = useNavigate();
+  const { address } = useAuth();
+  const { signTransaction } = useWallet();
+
+  const handlePurchase = useCallback(async (datasetId: string) => {
+    if (!datasetId) throw new Error('Item hash is undefined');
+    if (!signTransaction) throw new Error('Wallet not connected');
+
+    const { transaction } = await createTransaction({
+      signer: address,
+      datasetId
+    })
+
+    const serializedBuffer = Buffer.from(transaction, 'base64');
+    const unsignedTransaction = VersionedTransaction.deserialize(serializedBuffer);
+    const signedTransaction = await signTransaction(unsignedTransaction);
+    const signedSerializedBase64 = Buffer.from(signedTransaction.serialize()).toString('base64');
+    const response = await sendTransaction({
+      transaction: signedSerializedBase64,
+      datasetId,
+    });
+
+    console.log(response);
+
+    navigate(`/data/${datasetId}`);
+  }, [address, signTransaction, navigate]);
+
+  return { handlePurchase };
 };
